Finished developing movie details view controller and implemented UICollectionViewFlowLayout for advanced layouts

+++ Movieroo/Modules/Movie Detail/MovieReview.swift
@@ -40,7 +40,7 @@ struct Review: Codable {
 struct AuthorDetails: Codable {
     let name, username: String
     let avatarPath: String?
-    let rating: Int?
+    let rating: Double?
 
     enum CodingKeys: String, CodingKey {
         case name, username
